requirements_server.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.5
billiard==4.2.1
bleach==5.0.1
boto3==1.39.1
botocore==1.39.1
celery==5.5.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2025.6
cryptography==44.0.0
defusedxml==0.7.1
diff-match-patch
Django==5.2.4
django-allauth==65.9.0
django-bleach==3.1.0
django-channels==0.7.0
django-ckeditor==6.7.3
django-crispy-forms==2.4
django-extensions==3.2.3
django-intl-tel-input==0.4.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-phonenumber-field==8.1.0
django-quill-editor==0.1.42
django-recaptcha==4.1.0
django-storages==1.14.6
djangorestframework==3.16.1
environs==14.2.0
Faker==37.4.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.5.3
MarkupSafe
marshmallow==3.22.0
oauthlib==3.2.2
packaging==24.2
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python3-openid==3.2.0
pytz==2024.2
quill-delta==1.0.3
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.0
shortuuid==1.0.11
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stripe==11.6.0
tinycss2==1.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
